+++ src/test/java/org/plumelib/StringsPlumeTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Locale;
 import org.checkerframework.common.value.qual.ArrayLen;
 import org.junit.jupiter.api.Test;